Run dispatcher-dependent navigation test on JVM only
