+++ sample/shared/src/commonTest/kotlin/dev/chrisbanes/haze/sample/SamplesTest.kt
@@ -3,11 +3,7 @@
 
 package dev.chrisbanes.haze.sample
 
-import androidx.compose.material3.Button
 import androidx.compose.material3.Text
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.test.ExperimentalTestApi
@@ -23,34 +19,9 @@ import assertk.assertions.contains
 import assertk.assertions.isEqualTo
 import dev.chrisbanes.haze.test.ContextTest
 import kotlin.test.Test
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.withContext
 
 @OptIn(ExperimentalTestApi::class)
 class SamplesTest : ContextTest() {
-  @Test
-  fun directLink_seedsNormalBackHistoryAndDoesNotReopenOnRecomposition() = runComposeUiTest {
-    var appTitle by mutableStateOf("Haze Samples")
-    val sample = Sample(route = "probe", title = "Probe", effects = listOf(SampleEffect.Glass)) { nav, effect ->
-      Button(onClick = { nav.popBackStack() }) { Text("Exit ${effect.label} probe") }
-    }
-    val selection = resolveSampleLaunch(mapOf("sample" to "probe", "effect" to "glass"), listOf(sample))
-      as SampleLaunchRequest.Selected
-    setContent {
-      Samples(appTitle = appTitle, samples = listOf(sample), initialSelection = selection)
-    }
-
-    onNodeWithText("Exit Glass probe").assertIsDisplayed()
-    withContext(Dispatchers.Main) { onNodeWithText("Exit Glass probe").performClick() }
-    onNodeWithTag("sample_list").assertIsDisplayed()
-    onNodeWithText("Exit Glass probe").assertDoesNotExist()
-    runOnIdle { appTitle = "Updated samples" }
-    onNodeWithText("Exit Glass probe").assertDoesNotExist()
-    withContext(Dispatchers.Main) { onNodeWithContentDescription("Back").performClick() }
-    onNodeWithTag("sample_effect_glass").assertIsDisplayed()
-    onNodeWithTag("sample_effect_blur").assertIsDisplayed()
-  }
-
   @Test
   fun commonSamples_routesAndTitlesAreUnique() {
     assertThat(CommonSamples.map(Sample::route).distinct().size)
